HVAC controller and model are essentially done
The only issue that remains is connecting the functions to the xml file which I am sure that David will be able to help us get that fixed and handled.

+++ AndroidStudio/VehicleInfotainmentSystem/app/src/main/java/Model/HVAC_Model.java
@@ -13,7 +13,7 @@ public class HVAC_Model {
         return temperature;
     }
 
-    public void setTemperature(byte temperature) {
+    public void setTemperature(int temperature) {
         if(this.temperature>90){this.temperature = 90;}
         else if (this.temperature<50){this.temperature = 50;}
     }
